move configuration validation to entire AssetConfig trait
