+++ Engine/Source/assets/src/assets/model.rs
@@ -1,11 +1,11 @@
-use crate::{Asset, AssetHandle, AssetType, Error, Result, assets::AssetConfig};
+use crate::{Asset, AssetHandle, AssetType, Error, Metadata, Result, assets::AssetConfig};
 
 use serde::{Deserialize, Serialize};
 
 use anyhow::Context;
 
 /// Type to configure a model.
-#[derive(Serialize, Deserialize, Clone, AssetConfig)]
+#[derive(Serialize, Deserialize, Clone)]
 pub struct ModelConfig {
     /// Path to .gltf. Relative to asset dir
     pub gltf: String,
@@ -13,6 +13,29 @@ pub struct ModelConfig {
 
 /// Raw glTF bytes for a model asset. The renderer is responsible for
 /// uploading this to the GPU after calling [`Handle::consume`].
+impl AssetConfig for ModelConfig {
+    /// Validates that the glTF file actually exists on disk.
+    ///
+    /// Resolves the path as `handle.dir().join(&self.gltf)` — i.e. relative
+    /// to the `.dirkasset` file's own directory — and checks [`Path::exists`].
+    ///
+    /// Logs a warning and returns `false` if the file is absent.
+    ///
+    /// [`Path::exists`]: std::path::Path::exists
+    fn validate(&self, meta: &Metadata) -> bool {
+        let path = meta.handle.dir().join(&self.gltf);
+        if !path.exists() {
+            warn!(
+                "asset {}: glTF file not found at '{}'",
+                meta.handle.raw(),
+                path.display()
+            );
+            return false;
+        }
+        true
+    }
+}
+
 #[derive(Clone)]
 pub struct ModelData {
     pub gltf: gltf::Document,
